third_party/WebKit/LayoutTests/bluetooth/requestDevice/same-device.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/testdriver.js"></script>
<script src="../../resources/testdriver-vendor.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../external/wpt/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  let devices = [];
  let push = device => devices.push(device);

  return setUpHealthThermometerAndHeartRateDevices()
    .then(() => requestDeviceWithTrustedClick({
      filters: [{services: [heart_rate.alias]}]}))
    .then(push)
    .then(() => requestDeviceWithTrustedClick({
      filters: [{services: [heart_rate.name]}]}))
    .then(push)
    .then(() => requestDeviceWithTrustedClick({
      filters: [{services: [heart_rate.uuid]}]}))
    .then(push)
    .then(() => {
    assert_equals(devices[0], devices[1]);
    assert_equals(devices[1], devices[2]);
  });
}, 'Returned device should always be the same.');
</script>
